+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 1.2 (pre-release):
+* [Enhancement #260] Add a setting to control the couple state on link.
 * [Change] Major update to the basic renderer module to increase UX experience.
 * [Fix #252] NRE when entering the vessel from EVA.